:recycle: refactor(user): CurrentMemberResolver owner 되도록 수정
