+++ src/main/java/com/example/fitpassserver/global/common/support/LoginUserFinder.java
@@ -9,7 +9,5 @@ public interface LoginUserFinder {
 
     boolean existsByLoginId(String loginId);
 
-    boolean existsByName(String name);
-
     boolean existsByPhoneNumber(String phoneNumber);
 }
